requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
contourpy==1.1.1
cycler==0.12.1
docopt==0.6.2
efficientnet_pytorch==0.6.3
et_xmlfile==2.0.0
fonttools==4.57.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
importlib_resources==6.4.5
kiwisolver==1.4.7
kornia==0.5.4
Markdown==3.3.4
matplotlib==3.7.5
munch==2.5.0
numpy==1.20.3
oauthlib==3.1.1
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
Pillow==8.2.0
pipreqs==0.4.13
pretrainedmodels==0.7.4
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.1.1.post2209072238
timm==0.3.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.61.1
typing-extensions==3.10.0.0
tzdata==2025.3
urllib3==1.26.5
Werkzeug==2.0.1
yarg==0.1.10
zipp==3.20.2
